requirements.txt
Django==5.2.1
django-unfold==0.58.0
crispy-bootstrap5==2025.4
django-crispy-forms==2.4
python-dotenv==1.0.1
Pillow==11.2.1
django-widget-tweaks==1.5.0
